Workaround for `xcodebuild` v15 not resolving symlinks
`xcodebuild` as of version 15 fails when given a command line like:

```shell
xcodebuild -create-xcframework\
  -framework /var/folders/<elided>/osx-universal/publish/SampleKit.framework\
  -framework /var/folders/<elided>/iossimulator-universal/publish/SampleKit.framework\
  -framework /var/folders/<elided>/ios-arm64/publish/SampleKit.framework\
  -output /var/folders/<elided>/apple-universal/publish/SampleKit.xcframework
```

It returns exit code 70 and prints an error message similar to:

```
error: cannot compute path of binary
 'Path(str: "/private/var/folders/<elided>/osx-universal/publish/SampleKit.framework/Versions/A/
relative to that of '/var/folders/<elided>/osx-universal/publish/SampleKit.framework'
```

Note the mismatch between `/private/var` and `/var` -- on OSX, `/var` is a symlink to `/private/var`.

`xcodebuild` v15 seems to resolve symlinks for some arguments but not others and ends up confused. Unfortunately, .NET APIs do not resolve symlinks at the *start* of a path, only for the *last* component.

Hence, this helper resolves symlinks in the initial `/var` path segment explicitly, as a workaround until `xcodebuild` is updated to work correctly with symlinked paths again.

Local testing confirmed this approach works around the problem for now.