Try falling back to `/packs` dir relative to `/etc/dotnet/install_location` path
On Linux, the official .NET install script does not create the marker file at `/etc/dotnet/install_location[_arm64]`, and the .NET tools are installed into `~/.dotnet` by default.

This means that fallback to a hardcoded location is unlikely to work for a preview release, e.g. 9.0-rc1.

If the user is able to manually create `/etc/dotnet/install_location`, we now pick that path up, and try locating the `/packs` dir relative to it.

If this also fails, we fall back to the pre-existing hardcoded full path, as before.