Add `NoWarn` to the configuration file
Allow suppressing specific warnings occurring in the generated code.
This may come in handy for generic C# compiler or analyzer warnings.

But as of .NET 10, if the generated code happens to reach any
[preview or experimental .NET APIs](https://learn.microsoft.com/dotnet/fundamentals/runtime-libraries/preview-apis#experimentalattribute),
the ability to suppress warnings becomes *required* to continue.

(e.g. `SYSLIB5006` must be suppressed to use a few new `System.Security.Cryptography` types added in .NET 10)

Document the `NoWarn` option and add it to the config files samples in this repo for one for one known use-case.