Retarget to  .NET 10 (#104)
* Retarget to  .NET 10

* Update readme for .NET 10

* Update Swift test for .NET 10

* Enable experimental .NET APIs in generated projects

New in .NET 10 are compilation errors similar to:

> error SYSLIB5006: 'System.Security.Cryptography.SlhDsaAlgorithm' is for evaluation purposes only and is subject to change or removal in future updates.
> Suppress this diagnostic to proceed. (https://aka.ms/dotnet-warnings/SYSLIB5006)

The generated code reflects over and ends up including APIs such as this one.
Allow the generated project to use experimental APIs to resolve this issue.

ref. https://learn.microsoft.com/dotnet/fundamentals/runtime-libraries/preview-apis

> When building against an experimental API, the compiler produces an error. Each feature that's marked as experimental has its own separate diagnostic ID. To express consent to using an experimental API, you suppress the specific diagnostic. You can do that via any of the means for suppressing diagnostics, but the recommended way is to add the diagnostic to the project's `<NoWarn>` property.

`SYSLIB5006` added here seems to suffice in addressing current use cases;
however this opt-in may need to be extended further in the future ref:

> Since each experimental feature has a separate ID, consenting to using one experimental feature doesn't consent to using another.

* Add `NoWarn` to the configuration file

Allow suppressing specific warnings occurring in the generated code.
This may come in handy for generic C# compiler or analyzer warnings.

But as of .NET 10, if the generated code happens to reach any
[preview or experimental .NET APIs](https://learn.microsoft.com/dotnet/fundamentals/runtime-libraries/preview-apis#experimentalattribute),
the ability to suppress warnings becomes *required* to continue.

(e.g. `SYSLIB5006` must be suppressed to use a few new `System.Security.Cryptography` types added in .NET 10)

Document the `NoWarn` option and add it to the config files samples in this repo for one for one known use-case.

* Move `NoWarn` config under `Build`

ref. https://github.com/royalapplications/beyondnet/pull/104#issuecomment-3184606699